### 2018-07-13 #823 @whedon
[PRE REVIEW]: sdmbench: An R Package for Benchmarking Species Distribution Models
**Submitting author:** @boyanangelov (<a href="http://orcid.org/0000-0001-5068-4234">Boyan Angelov</a>)
**Repository:** <a href="https://github.com/boyanangelov/sdmbench" target ="_blank">https://github.com/boyanangelov/sdmbench</a>
**Version:** v0.1.2
