### 2018-08-01 boyanangelov/sdmbench#1 @goldingn
R version dependency required?
Hi Boyan, I'm Nick, one of the [reviewers of sdmbench for JOSS](https://github.com/openjournals/joss-reviews/issues/847). I'm starting my code review now, so will post issues as they come up, then I'll summarise them in the review thread later.

---

I just went to install the package following the readme instructions:
```r
devtools::install_github("boyanangelov/sdmbench")
```
and it failed to install with the error:
> ```
> ERROR: this R is version 3.4.3, package 'sdmbench' requires R >= 3.4.4
> ```
which is because there's a dependency on a very recent version of R in [DESCRIPTION](https://github.com/boyanangelov/sdmbench/blob/master/DESCRIPTION#L14).

Is that dependency on a super-recent R version really necessary?

If not, it might be worth removing this entry altogether.

If it is necessary, it would be worth mentioning that in the installation information.
