### 2018-08-01 #847 @goldingn
## Software paper review

Hi @boyanangelov, here's a short review of the software paper:

I think it reads really well. The statements of the problem that SDM solves, and the issue of benchmarking SDMs that sdmbench solves, are nice and succinct. The rest of the document does a nice job of explaining what the package provides that others don't. Great work!

### Affiliation
You don't have an affiliation listed. Was that deliberate?

### Citations

One of the citations has as the author: "Datasets, Healthcare Mimic", which looks like a bibtex malformation of the author. You should be able to change the author to: "Healthcare MIMIC Datasets" (which I'm guessing you were after) by editing the bibtex entry to have `{}`s around the author, like [this](http://timmurphy.org/2011/02/15/using-an-organisation-name-as-an-author-in-bibtex/). Though there's also [this version of the paper](https://arxiv.org/abs/1710.08531), with a human name for the author, so maybe that's worth using instead.

The citation for Wu et al. doesn't have a date. Maybe it would be worth citing [this dated arxiv paper](https://arxiv.org/abs/1703.00564) instead.

The ENMeval citation has smallcaps html tags for Maxent, that didn't render properly in the bibliography. It's probably possible to get them working in the file (e.g. [here](https://tex.stackexchange.com/questions/191401/bold-or-smallcaps-typesetting-in-bibliography)?), or you could just omit the small caps, as in the [Google Scholar bibtex version](https://scholar.googleusercontent.com/scholar.bib?q=info:gIvEknnKvx4J:scholar.google.com/&output=citation&scisig=AAGBfm0AAAAAW2ETw1iIMFXKPWMo9zgDRU1bn9ZJqFTI&scisf=4&ct=citation&cd=-1&hl=en).
