### 2018-08-01 79f9c48
Fix citation entries
