+++ paper/paper.bib
@@ -67,41 +67,52 @@ year = {2016}
 }
 
 @article{Wu2017,
-archivePrefix = {arXiv},
-arxivId = {arXiv:1703.00564v2},
-author = {Wu, Zhenqin and Ramsundar, Bharath and Feinberg, Evan N and Gomes, Joseph and Geniesse, Caleb and Pappu, Aneesh S and Leswing, Karl and Pande, Vijay},
-eprint = {arXiv:1703.00564v2},
-title = {{MoleculeNet: A Benchmark for Molecular Machine Learning}}
+  author    = {Zhenqin Wu and
+               Bharath Ramsundar and
+               Evan N. Feinberg and
+               Joseph Gomes and
+               Caleb Geniesse and
+               Aneesh S. Pappu and
+               Karl Leswing and
+               Vijay S. Pande},
+  title     = {MoleculeNet: {A} Benchmark for Molecular Machine Learning},
+  journal   = {CoRR},
+  volume    = {abs/1703.00564},
+  year      = {2017},
+  url       = {http://arxiv.org/abs/1703.00564},
+  archivePrefix = {arXiv},
+  eprint    = {1703.00564},
+  timestamp = {Wed, 07 Jun 2017 14:43:02 +0200},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/WuRFGGPLP17},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
 @article{Purushotham2017,
-archivePrefix = {arXiv},
-arxivId = {arXiv:1710.08531v1},
-author = {Datasets, Healthcare Mimic},
-eprint = {arXiv:1710.08531v1},
-keywords = {deep learning models,icd-9 code group prediction,length of stay,mortality prediction,super learner algorithm},
-title = {{Benchmark of Deep Learning Models on Large Healthcare MIMIC Datasets}},
-year = {2017}
+  author    = {Sanjay Purushotham and
+               Chuizheng Meng and
+               Zhengping Che and
+               Yan Liu},
+  title     = {Benchmark of Deep Learning Models on Large Healthcare {MIMIC} Datasets},
+  journal   = {CoRR},
+  volume    = {abs/1710.08531},
+  year      = {2017},
+  url       = {http://arxiv.org/abs/1710.08531},
+  archivePrefix = {arXiv},
+  eprint    = {1710.08531},
+  timestamp = {Thu, 02 Nov 2017 14:25:36 +0100},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1710-08531},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
 @article{Muscarella2014,
[…]
