+++ paper/paper.md
@@ -8,10 +8,6 @@ tags:
 authors:
   - name: Boyan Angelov
     orcid: 0000-0001-5068-4234
-    affiliation:
-affiliations:
- - name:
-   index: 1
 date: 11 July 2018
 bibliography: paper.bib
 ---
