### 2018-08-01 c6a7ec6
Remove R version dependency
