+++ DESCRIPTION
@@ -11,7 +11,6 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
-Depends: R (>= 3.4.4)
 Imports:
     dismo,
     dplyr,
