### 2018-08-01 85e8688
Improve style of shiny server file
