+++ inst/shiny/server.r
@@ -12,36 +12,38 @@ server <- function(input, output) {
     observeEvent(input$go, {
         req(input$text)
 
-
         progress <- shiny::Progress$new()
 
         progress$set(message = "Status", value = 0)
-        progress$inc(1/4, detail = "Downloading species occurence data and climate variables...")
+        progress$inc(1/4,
+            detail = "Downloading species occurence data and climate
+            variables...")
 
-        benchmarking_data <- get_benchmarking_data(scientific_name = input$text,
-                                                   limit = input$limit,
-                                                   climate_type = input$climate_type,
-                                                   climate_resolution = input$climate_resolution,
-                                                   projected_model = input$projected_model,
-                                                   rcp = input$rcp,
-                                                   year = input$years)
+        benchmarking_data <- get_benchmarking_data(
+            scientific_name = input$text,
+            limit = input$limit,
+            climate_type = input$climate_type,
+            climate_resolution = input$climate_resolution,
+            projected_model = input$projected_model,
+            rcp = input$rcp,
+            year = input$years)
 
         progress$inc(1/4, detail = "Processing data...")
 
-        benchmarking_data$df_data <- partition_data(dataset_raster = benchmarking_data$raster_data,
-                                                   dataset = benchmarking_data$df_data,
-                                                   env = benchmarking_data$raster_data$climate_variables,
-                                                   method = input$data_partitioning_type)
+        benchmarking_data$df_data <- partition_data(
+            dataset_raster = benchmarking_data$raster_data,
+            dataset = benchmarking_data$df_data,
+            env = benchmarking_data$raster_data$climate_variables,
+            method = input$data_partitioning_type)
 
         progress$inc(1/4, detail = "Plotting map...")
-
 
         output$table <- renderTable(head(benchmarking_data$df_data))
         output$species_name <- renderText(input$text)
 
-
         output$occ_map <- leaflet::renderLeaflet({
-            leaflet::leaflet(data = benchmarking_data$raster_data$coords_presence) %>%
+            leaflet::leaflet(
+                data = benchmarking_data$raster_data$coords_presence) %>%
                 leaflet::addTiles() %>%
                 leaflet::addCircleMarkers(~x, ~y, fillOpacity = 0.3)
         })
@@ -50,7 +52,7 @@ server <- function(input, output) {
         progress$close()
 
         v$data <- benchmarking_data
-        v$partitioning_type = input$data_partitioning_type
+        v$partitioning_type <- input$data_partitioning_type
 
         shinyjs::enable("go_bmr")
         shinyjs::enable("go_maxent")
@@ -63,15 +65,15 @@ server <- function(input, output) {
         benchmarking_data <- v$data
         partitioning_type <- v$partitioning_type
 
-
         progress_bmr <- shiny::Progress$new()
         progress_bmr$set(message = "Status", value = 0)
         progress_bmr$inc(1/3, detail = "Benchmarking...")
 
-        learners_list = input$checkGroup
+        learners_list <- input$checkGroup
         learners <- list()
-        for (i in 1:length(learners_list)) {
-            learners[[i]] <- mlr::makeLearner(learners_list[[i]], predict.type = "prob")
+        for (i in seq_along(learners_list)) {
+            learners[[i]] <- mlr::makeLearner(learners_list[[i]],
+                                              predict.type = "prob")
         }
 
         benchmarking_data$df_data <- na.omit(benchmarking_data$df_data)
@@ -84,63 +86,78 @@ server <- function(input, output) {
         output$bmr_plot1 <- renderPlot(plotBMRBoxplots(bmr, measure = mlr::auc))
[…]
