### 2018-08-01 1b89e42
Add package-level helpfile
