### 2018-08-01 #847 @boyanangelov
Hi @goldingn,

Thank you for the positive opinion on the paper. About the affiliation - yes, at the moment I am not formally academically affiliated. I work as a machine learning researcher in the industry but in a different field.

I pushed the citation changes to Github, I hope they are fixed now. Thanks again for taking the time to review this package!
