### 2018-08-01 boyanangelov/sdmbench#1 @boyanangelov
Hello Nick, thank you for the suggestion! I decided to remove it. Hopefully now the installation becomes more straightforward.
