### 2018-08-07 #847 @karthik
@boyanangelov Can you use a former institution if any part of this work was begun while employed there? If not, you should just use your current employer as the affiliation.
