### 2018-08-08 #847 @boyanangelov
@karthik Thank you for the suggestions, I just added my current employer as the affiliation.
