### 2018-08-16 boyanangelov/sdmbench#5 @sckott
Consider rgbif::occ_data?
from JOSS review

I noticed that in `get_benchmarking_data()` you use `rgbif::occ_search`. You might consider `rgbif::occ_data` as it has all the same user interface (parameters) but only spends time internally getting and parsing the occurrence data and completely drops the other data that you don't want.

Also note that occ_search/occ_data use the GBIF occurrence API, which is limited to 200K results for any one query. So if you know or think you might need more than that you should use the download API via `rgbif::occ_download` and friends

also also note that we now have an interface to the GBIF maps API via `rgbif::map_fetch` which is a super fast way to get a raster map of occurrences, so you're not getting the actual occurrence data but a quick summary of the data- see https://www.gbif.org/developer/maps - Don't know if this is appropriate or not for your use case(s) but worth knowing about
