### 2018-08-16 #847 @sckott
### software

- had an issue with the vignette building locally for me, and examples seemed to be not in sync with ([issue 4](https://github.com/boyanangelov/sdmbench/issues/4)) - it's not quite clear to me if the rendered vignette online is behind the v0.1.2 tagged version or ahead?
- some notes on GBIF data ([issue 5](https://github.com/boyanangelov/sdmbench/issues/5))
- For __Statement of Need__, i'd say that's partially fulfilled. the problem is clearly stated but i don't see a clear statment of who the software is targeted at
- For __Installation Instructions__, i think the github repo (Readme, vignette) could use more instructions given that it depends on Python and Java and Maxent (a separate standalone piece of software), like what to do if users are having trouble with a python dependency or virtualenv or python version, etc.
- For __Functionality documentation__, as someone not super familiar with SDM's and macine learning, i'd love more documentation in each fxns man page about what possible values are for each parameter (if there's a finite set), etc.
- i'd pull the Code of Conduct out of the contributing file into a separate file `CODE_OF_CONDUCT.md`

### paper

- Wu et al. has a DOI: https://doi.org/10.1039/C7SC02664A
- affilitation not present for author, already pointed out by other reviewer
- author does a nice job introducing the problem and defining how this effort is different from previous art
- "The data processing pipeline relies on external functions" maybe is more accurate as "software" or "packages" cause in theory a function doesn't need to be in a package :)
