### 2018-08-21 1b19289
Replace occ_search with the more efficient get_data; change input max records number accordingly
