+++ inst/shiny/ui.r
@@ -11,7 +11,7 @@ dashboardPage(skin = "green",
         ),
         tags$hr(),
         textInput("text", "Species name:"),
-                numericInput("limit", "Max number of records:", 1000, min = 1, max = 15000),
+                numericInput("limit", "Max number of records:", 1000, min = 1, max = 200000),
                 selectInput("climate_type", "Climate type:",
                             c("Default" = "default",
                               "Future" = "future")),
