### 2018-08-21 boyanangelov/sdmbench#5 @boyanangelov
Hello @sckott, thanks for reviewing! I replaced the call to `occ_search` with `occ_data`. `occ_download` and `map_fetch` are nice. I think for this release the existing functions should do, but will keep those in mind for a future one.
