### 2018-08-21 c9cc01b
Add Wu2017 DOI; minor paper text edit
