+++ paper/paper.bib
@@ -75,6 +75,7 @@ year = {2016}
                Aneesh S. Pappu and
                Karl Leswing and
                Vijay S. Pande},
+  doi       = {10.1039/C7SC02664A},
   title     = {MoleculeNet: {A} Benchmark for Molecular Machine Learning},
   journal   = {CoRR},
   volume    = {abs/1703.00564},
