### 2018-08-21 1606d9f
Move code of conduct to a separate file
