### 2018-08-21 7f666d1
Improve README by adding package target audience
