### 2018-08-21 28130df
Add Muscarella2014 DOI; improve bib formatting
