+++ paper/paper.bib
@@ -106,30 +106,31 @@ year = {2016}
 }
 
 @article{Muscarella2014,
-  title={ENM eval: An R package for conducting spatially independent evaluations and estimating optimal model complexity for Maxent ecological niche models},
-  author={Muscarella, Robert and Galante, Peter J and Soley-Guardia, Mariano and Boria, Robert A and Kass, Jamie M and Uriarte, Mar{\'\i}a and Anderson, Robert P},
-  journal={Methods in Ecology and Evolution},
-  volume={5},
-  number={11},
-  pages={1198--1205},
-  year={2014},
-  publisher={Wiley Online Library}
+  title     = {ENM eval: An R package for conducting spatially independent evaluations and estimating optimal model complexity for Maxent ecological niche models},
+  author    = {Muscarella, Robert and Galante, Peter J and Soley-Guardia, Mariano and Boria, Robert A and Kass, Jamie M and Uriarte, Mar{\'\i}a and Anderson, Robert P},
+  doi       = {10.1111/2041-210X.12261},
+  journal   = {Methods in Ecology and Evolution},
+  volume    = {5},
+  number    = {11},
+  pages     = {1198--1205},
+  year      = {2014},
+  publisher = {Wiley Online Library}
 }
 
 @misc{Rochette2017,
-  author       = {Sébastien Rochette},
-  title        = {{SDMSelect: A R-package for cross-validation model
-                   selection and species distribution mapping}},
-  month        = sep,
-  year         = 2017,
-  doi          = {10.5281/zenodo.894344},
-  url          = {https://doi.org/10.5281/zenodo.894344}
+  author    = {Sébastien Rochette},
+  title     = {SDMSelect: A R-package for cross-validation model
+                selection and species distribution mapping},
+  month     = sep,
+  year      = 2017,
+  doi       = {10.5281/zenodo.894344},
+  url       = {https://doi.org/10.5281/zenodo.894344}
 }
 
 @Manual{scrubr,
-    title = {scrubr: Clean Biological Occurrence Records},
-    author = {Scott Chamberlain},
-    year = {2016},
-    note = {R package version 0.1.1},
-    url = {https://CRAN.R-project.org/package=scrubr},
+    title   = {scrubr: Clean Biological Occurrence Records},
+    author  = {Scott Chamberlain},
+    year    = {2016},
+    note    = {R package version 0.1.1},
+    url     = {https://CRAN.R-project.org/package=scrubr},
   }
