### 2018-08-21 f338d47
Add additional documentation for Python and Java installations
