### 2018-08-21 boyanangelov/sdmbench#4 @boyanangelov
Hello @sckott. Thank you for submitting the issue! I did not manage to replicate the problem, but have added additional information about the Python and Java installations to both the `README.md` and the package vignette. If there is a working installation of Python, followed by the `keras` setup in the documentation, the vignette should now compile successfuly. Could you let me know if this is the right level of detail, and if it is I can close the issue?

I also updated the uploaded vignette to be in line with the version of the package, they will be in sync from now on.
