### 2018-08-21 c6bc6e6
Improve function documentation by adding additional information on possible parameter values
