### 2018-08-27 #847 @goldingn
@boyanangelov sorry for being so slow with the code review. I've been away for a few weeks, but have tomorrow afternoon blocked out to get my comments to you!
