### 2018-08-28 boyanangelov/sdmbench#4 @goldingn
I'm having a different problem installing whilst building the vignette:
```r
devtools::install_github("boyanangelov/sdmbench", force = TRUE, build_vignettes = TRUE)
```
```
Warning in predict.lm(object, newdata, se.fit, scale = 1, type = ifelse(type ==  :
  prediction from a rank-deficient fit may be misleading
  ...
Quitting from lines 246-252 (sdmbench_vignette.Rmd)
Error: processing vignette 'sdmbench_vignette.Rmd' failed with diagnostics:
attempt to select less than one element in get1index
Execution halted
Installation failed: Command failed (1)
```

I have a working installation of maxent. E.g. I can do:
```r
library(dismo)
example(maxent)
```
which completes successfully.

Installing without vignettes, and then running the vignette as an R notebook, I get the following error in the chunk mentioned above:
```
Error in eval@models[[best_model_id]] : attempt to select less than one element in get1index
```
traceback:
```
4.  .local(object, ...)
3.  dismo::predict(raster_data$climate_variables, eval@models[[best_model_id]])
2.  dismo::predict(raster_data$climate_variables, eval@models[[best_model_id]])
1.  evaluate_maxent(benchmarking_data$raster_data, method = "block")
```
