### 2018-08-28 #847 @goldingn
Wow, thanks @whedon! I for one welcome ... etc.
