### 2018-08-28 boyanangelov/sdmbench#4 @goldingn
OK, looks like my error is actually due to an API change in ENMeval.

On [this line](https://github.com/boyanangelov/sdmbench/blob/master/R/evaluate_maxent.r#L20) you access the `full.AUC` column of the results dataframe. I have version 0.3.0 (which hit CRAN 2 weeks ago), which doesn't have that column, and the Value section of the helpfile says for the results dataframe:
> Note that the names of some columns changed as of Version 0.3.0.

I would suggest you change the code and add ENMeval v0.3.0 as a version dependency. Alternatively, you could write code to handle whatever version of ENMeval is installed, but you'd have to be confident nothing else is different between the versions, and that could be tricky to maintain.

I'm having some curious trouble trying to install an earlier version of ENMeval, so will wait for you to fix this before trying again.
