### 2018-08-28 e294f8f
Fix typos in function documentation
