### 2018-08-28 84a7241
Fix breaking ENMEval AUC extraction
