### 2018-08-28 boyanangelov/sdmbench#4 @boyanangelov
Thanks for going so deep to find the exact issue @goldingn . I adjusted the code to use the ENMEval v0.3.0 AUC extraction, and also added that as a dependency. Can you try to build the vignette again?
