### 2018-08-28 0601b0d
Use new version of roxygen
