+++ DESCRIPTION
@@ -10,7 +10,7 @@ Description: This package provides tools and functions to benchmark Species Dist
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
 Imports:
     dismo,
     dplyr,
