### 2018-08-28 ce21a0f
Fix links for package-level helpfile
