### 2018-08-28 c95df3b
Add improved function documentation examples
