+++ R/plot_sdm_map.r
@@ -8,9 +8,34 @@
 #'
 #' @return An interactive leaflet map, showing the species distribution.
 #' @examples
-#' \dontrun{
-#' plot_sdm_map(raster_data = benchmarking_data$raster_data$climate_variables, bmr_models = bmr_models, model_id = best_results$learner.id[4], model_iteration = best_results$iter[4], map_type = "static")
-#' }
+#' # download benchmarking data
+#' benchmarking_data <- get_benchmarking_data("Lynx lynx",
+#'                                            limit = 1500,
+#'                                            climate_resolution = 10)
+#'
+#' # create a list of learners (algorithms) to compare
+#' learners <- list(mlr::makeLearner("classif.randomForest",
+#'                                   predict.type = "prob"),
+#'                  mlr::makeLearner("classif.logreg",
+#'                                   predict.type = "prob"))
+#'
+#' # run the benchmarking
+#' bmr <- benchmark_sdm(benchmarking_data$df_data,
+#'                      learners = learners,
+#'                      dataset_type = "default",
+#'                      sample = FALSE)
+#'
+#' # get best model results
+#' # you should obtain a dataframe containing the highest performing (by AUC)
+#' # algorithm name, iteration and associated AUC
+#' best_results <- get_best_model_results(bmr)
+#'
+#' # plot the SDM map of the best performing model
+#' plot_sdm_map(raster_data = benchmarking_data$raster_data$climate_variables,
+#'              bmr_models = bmr$learners,
+#'              model_id = best_results$learner.id[1],
+#'              model_iteration = best_results$iter[1],
+#'              map_type = "static")
 plot_sdm_map <- function(raster_data, bmr_models, model_id, model_iteration, map_type = "static") {
     if (model_id == "classif.logreg") {
         model <- bmr_models$benchmarking_data[[model_id]][[model_iteration]]$learner.model
