### 2018-08-28 c632a03
Add improved function documentation examples
