+++ man/evaluate_maxent.Rd
@@ -18,8 +18,13 @@ A list containing AUC value and predict object (for plotting).
 Evaluate MaxEnt model performance
 }
 \examples{
-\dontrun{
-benchmarking_data <- get_benchmarking_data("Lynx lynx", limit = 1500, climate_resolution = 10)
+# download benchmarking data
+benchmarking_data <- get_benchmarking_data("Lynx lynx",
+                                           limit = 1500,
+                                           climate_resolution = 10)
+# run MaxEnt evaluation
 maxent_results <- evaluate_maxent(benchmarking_data$raster_data, method = "block")
-}
+
+# get AUC of best model run
+maxent_results$best_auc
 }

[…]
