### 2018-08-29 08e86f1
Stop running of examples for travis ci
