+++ R/evaluate_dl.r
@@ -6,6 +6,7 @@
 #'
 #' @return A tibble containing the model estimates.
 #' @examples
+#' \dontrun{
 #' # download benchmarking data
 #' benchmarking_data <- get_benchmarking_data("Lynx lynx",
 #'                                            limit = 1500,
@@ -37,6 +38,7 @@
 #' # create evaluation tibble containing deep learning training results
 #' keras_evaluation <- evaluate_dl(keras_results$model, benchmarking_data_dl)
 #' head(keras_evaluation)
+#' }
 evaluate_dl <- function(model_keras, input_data) {
     yhat_keras_class_vec <- keras::predict_classes(object = model_keras,
                                                    x = as.matrix(input_data$test_tbl)) %>%
