+++ R/partition_data.r
@@ -8,6 +8,7 @@
 #'
 #' @return A dataframe partitioned using the selected method.
 #' @examples
+#' \dontrun{
 #' # download benchmarking data
 #' benchmarking_data <- get_benchmarking_data("Lynx lynx",
 #'                                            limit = 1500,
@@ -35,6 +36,7 @@
 #'
 #' # you can perform a sanity check here as well, you should see two groups
 #' table(benchmarking_data$df_data$label)
+#' }
 partition_data <- function(dataset_raster, dataset, env, method) {
     if (method == "default") {
         result_dataset <- dataset
