### 2018-08-29 9605591
Fix DL test
