### 2018-08-31 boyanangelov/sdmbench#4 @goldingn
Seems to work for me now, thanks!

Not sure whether @sckott's issue was resolved though?
