### 2018-08-31 boyanangelov/sdmbench#4 @boyanangelov
Ok @goldingn . @sckott can you try building the vignette again (the package one is synced with the uploaded one now) ? I added more documentation for the python Keras stuff as well.
