### 2018-09-01 11c7cc8
Further improvements on documentation examples
