+++ R/plot_dl_map.r
@@ -10,8 +10,7 @@
 #' \dontrun{
 #' # download benchmarking data
 #' benchmarking_data <- get_benchmarking_data("Lynx lynx",
-#'                                            limit = 1500,
-#'                                            climate_resolution = 10)
+#'                                            limit = 1500)
 #'
 #' # transform benchmarking data into a format suitable for deep learning
 #' # if you have previously used a partitioning method you should specify it here
