+++ R/evaluate_dl.r
@@ -9,15 +9,14 @@
 #' \dontrun{
 #' # download benchmarking data
 #' benchmarking_data <- get_benchmarking_data("Lynx lynx",
-#'                                            limit = 1500,
-#'                                            climate_resolution = 10)
+#'                                            limit = 1500)
 #'
 #' # transform benchmarking data into a format suitable for deep learning
 #' # if you have previously used a partitioning method you should specify it here
 #' benchmarking_data_dl <- prepare_dl_data(input_data = benchmarking_data$df_data,
 #'                                        partitioning_type = "default")
 #'
-#' # perform sanity check on the transformed dataset
+#' # perform sanity checks on the transformed data
 #' # for the training set
 #' head(benchmarking_data_dl$train_tbl)
 #' table(benchmarking_data_dl$y_train_vec)
@@ -29,13 +28,13 @@
 #' # train neural network
 #' keras_results <- train_dl(benchmarking_data_dl)
 #'
-#' # inspect results
+#' # inspect training results
 #' keras_results$history
 #'
-#' # you can also plot the results
+#' # you can also plot them
 #' plot(keras_results$history)
 #'
-#' # create evaluation tibble containing deep learning training results
+#' # create evaluation tibble containing training results
 #' keras_evaluation <- evaluate_dl(keras_results$model, benchmarking_data_dl)
 #' head(keras_evaluation)
 #' }
