+++ man/train_dl.Rd
@@ -19,8 +19,7 @@ Train deep learning SDM
 \dontrun{
 # download benchmarking data
 benchmarking_data <- get_benchmarking_data("Lynx lynx",
-                                           limit = 1500,
-                                           climate_resolution = 10)
+                                           limit = 1500)
 
 # transform benchmarking data into a format suitable for deep learning
 # if you have previously used a partitioning method you should specify it here
@@ -42,7 +41,7 @@ keras_results <- train_dl(benchmarking_data_dl)
 # inspect results
 keras_results$history
 
-# you can also plot the results
+# you can also plot them
 plot(keras_results$history)
 }
 }

[…]
