### 2018-09-10 500078c
Undocumented code objects fix
