+++ R/run_sdmbench.r
@@ -8,6 +8,7 @@
 #' \dontrun{
 #' run_sdmbench()
 #' }
+#' @export
 run_sdmbench <- function() {
     app_path <- system.file("shiny", package = "sdmbench")
     return(shiny::runApp(app_path, launch.browser = TRUE))
