+++ R/plot_sdm_map.r
@@ -39,6 +39,7 @@
 #'              model_iteration = best_results$iter[1],
 #'              map_type = "static")
 #'}
+#'@export
 plot_sdm_map <- function(raster_data, bmr_models, model_id, model_iteration, map_type = "static") {
     if (model_id == "classif.logreg") {
         model <- bmr_models$benchmarking_data[[model_id]][[model_iteration]]$learner.model
