### 2018-09-10 #847 @boyanangelov
Thanks @goldingn. Just pushed the fixes for the undocumented functions.
