### 2018-09-22 ed9715e
Add explicit package referral for Shiny app - support for GalaxyE integration
