+++ inst/shiny/server.r
@@ -19,7 +19,7 @@ server <- function(input, output) {
             detail = "Downloading species occurence data and climate
             variables...")
 
-        benchmarking_data <- get_benchmarking_data(
+        benchmarking_data <- sdmbench::get_benchmarking_data(
             scientific_name = input$text,
             limit = input$limit,
             climate_type = input$climate_type,
@@ -30,7 +30,7 @@ server <- function(input, output) {
 
         progress$inc(1/4, detail = "Processing data...")
 
-        benchmarking_data$df_data <- partition_data(
+        benchmarking_data$df_data <- sdmbench::partition_data(
             dataset_raster = benchmarking_data$raster_data,
             dataset = benchmarking_data$df_data,
             env = benchmarking_data$raster_data$climate_variables,
@@ -77,83 +77,83 @@ server <- function(input, output) {
         }
 
         benchmarking_data$df_data <- na.omit(benchmarking_data$df_data)
-        bmr <- benchmark_sdm(benchmarking_data = benchmarking_data$df_data,
+        bmr <- sdmbench::benchmark_sdm(benchmarking_data = benchmarking_data$df_data,
                              learners = learners,
                              dataset_type =  partitioning_type,
                              sample = input$sample)
 
         progress_bmr$inc(1/3, detail = "Plotting benchmark results...")
         output$bmr_plot1 <- renderPlot(plotBMRBoxplots(bmr, measure = mlr::auc))
-        output$bmr_results <- renderTable(get_best_model_results(bmr))
+        output$bmr_results <- renderTable(sdmbench::get_best_model_results(bmr))
 
         bmr_models <- mlr::getBMRModels(bmr)
-        best_results <- get_best_model_results(bmr)
+        best_results <- sdmbench::get_best_model_results(bmr)
 
         output$model_map_1 <- leaflet::renderLeaflet(
-            plot_sdm_map(raster_data = benchmarking_data$raster_data,
+            sdmbench::plot_sdm_map(raster_data = benchmarking_data$raster_data,
                         bmr_models = bmr_models,
                         model_id = best_results$learner.id[1],
                         model_iteration = best_results$iter[1],
                         map_type = "interactive"))
 
         output$model_map_2 <- leaflet::renderLeaflet(
-            plot_sdm_map(raster_data = benchmarking_data$raster_data,
+            sdmbench::plot_sdm_map(raster_data = benchmarking_data$raster_data,
                         bmr_models = bmr_models,
                         model_id = best_results$learner.id[2],
                         model_iteration = best_results$iter[2],
                         map_type = "interactive"))
 
         output$model_map_3 <- leaflet::renderLeaflet(
-            plot_sdm_map(raster_data = benchmarking_data$raster_data,
+            sdmbench::plot_sdm_map(raster_data = benchmarking_data$raster_data,
                         bmr_models = bmr_models,
                         model_id = best_results$learner.id[3],
                         model_iteration = best_results$iter[3],
                         map_type = "interactive"))
 
         output$model_map_4 <- leaflet::renderLeaflet(
-            plot_sdm_map(raster_data = benchmarking_data$raster_data,
+            sdmbench::plot_sdm_map(raster_data = benchmarking_data$raster_data,
                         bmr_models = bmr_models,
                         model_id = best_results$learner.id[4],
                         model_iteration = best_results$iter[4],
                         map_type = "interactive"))
 
         output$model_map_5 <- leaflet::renderLeaflet(
-            plot_sdm_map(raster_data = benchmarking_data$raster_data,
+            sdmbench::plot_sdm_map(raster_data = benchmarking_data$raster_data,
                         bmr_models = bmr_models,
                         model_id = best_results$learner.id[5],
[…]
