### 2018-09-22 cb00187
Fix missing data in vignette bug
