### 2018-09-24 #847 @sckott
Happy with it now. I did open an issue https://github.com/boyanangelov/sdmbench/issues/9 when trying to run examples, but I imagine it's just an outdated readme issue or so
