### 2018-09-24 c4f01b8
Add missing steps to README
