+++ README.md
@@ -1,20 +1,34 @@
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Build Status](https://travis-ci.org/boyanangelov/sdmbench.svg?branch=master)](https://travis-ci.org/boyanangelov/sdmbench) [![DOI](https://zenodo.org/badge/138335161.svg)](https://zenodo.org/badge/latestdoi/138335161)
+[![Build
+Status](https://travis-ci.org/boyanangelov/sdmbench.svg?branch=master)](https://travis-ci.org/boyanangelov/sdmbench)
+[![DOI](https://zenodo.org/badge/138335161.svg)](https://zenodo.org/badge/latestdoi/138335161)
 
 sdmbench
 ========
 
 ![](logo.png)
 
-Species Distribution Modeling (SDM) is a field of increasing importance in ecology<sup>[1](#footnote1)</sup>. Several popular applications of SDMs are understanding climate change effects on species<sup>[2](#footnote2)</sup>, natural reserve planning<sup>[3](#footnote3)</sup> and invasive species monitoring<sup>[4](#footnote4)</sup>. The `sdmbench` package solves several issues related to the development and evaluation of those models by providing a consistent benchmarking workflow:
+Species Distribution Modeling (SDM) is a field of increasing importance
+in ecology<sup>[1](#footnote1)</sup>. Several popular applications of
+SDMs are understanding climate change effects on
+species<sup>[2](#footnote2)</sup>, natural reserve
+planning<sup>[3](#footnote3)</sup> and invasive species
+monitoring<sup>[4](#footnote4)</sup>. The `sdmbench` package solves
+several issues related to the development and evaluation of those models
+by providing a consistent benchmarking workflow:
 
 -   consistent species occurence data acquisition and preprocessing
--   consistent environmental data acquisition and preprocessing (both current data and future projections)
+-   consistent environmental data acquisition and preprocessing (both
+    current data and future projections)
 -   consistent spatial data partitioning
 -   integration of a wide variety of machine learning models
 -   graphical user interface for non/semi-technical users
 
-The end result of a `sdmbench` SDM analysis is to determine the model - data processing combination that results in the highest predictive power for the species of interest. Such an analysis is useful to researchers who want to avoid issues of model selection and evaluation, and want to rapidly test prototypes of species distribution models.
+The end result of a `sdmbench` SDM analysis is to determine the model -
+data processing combination that results in the highest predictive power
+for the species of interest. Such an analysis is useful to researchers
+who want to avoid issues of model selection and evaluation, and want to
+rapidly test prototypes of species distribution models.
 
 Installation
 ------------
@@ -24,25 +38,36 @@ Installation
 devtools::install_github("boyanangelov/sdmbench")
 ```
 
-There are several additional packages you need to install if you want to access the complete sdmbench functionality. First Tensorflow. You can use the `keras` package to install that (it is installed by the previous command). Note that this step requires a working Python installation on your system. Most modern operating systems have Python pre-installed, but if you are not sure you can check the [official website](https://www.python.org/).
+There are several additional packages you need to install if you want to
+access the complete sdmbench functionality. First Tensorflow. You can
+use the `keras` package to install that (it is installed by the previous
+command). Note that this step requires a working Python installation on
+your system. Most modern operating systems have Python pre-installed,
+but if you are not sure you can check the [official
[…]
