### 2018-09-24 boyanangelov/sdmbench#9 @boyanangelov
Just pushed a fix, some code was missing (a call to `partition_data`), it needs to be run before the `benchmark_sdm`. Now should be working!
