### 2018-09-24 #847 @boyanangelov
@sckott You are right, some code that was necessary as a previous step was missing from the `README`. I just pushed the changes.
