### 2018-09-24 bc1a52c
Add an additional line for newly added code to README
