### 2018-09-27 ac9d6f9
Add DOI for v0.1.3 to README and paper
