### 2018-09-27 #847 @boyanangelov
@karthik Great! I just made a release with all the JOSS fixes and improvements (v.0.1.3), archived in Zenodo (http://doi.org/10.5281/zenodo.1436376) and changed the DOI and version number in the paper.
