### 2018-09-29 #847 @boyanangelov
@sckott @goldingn Thank you for reviewing! The package improved greatly from your feedback. @karthik thank you for organizing everything, and @arfon thank you and everyone else at JOSS for this journal - I will be gladly recommending it to my colleagues!
